requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
fonttools==4.55.3
html5lib==1.1
idna==3.10
inflection==0.5.1
lxml==5.3.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
